Implement backend lifecycle management for destroy operations
- Refactor the DestroyFunc in down_test.go to return a boolean indicating success.
- Introduce a new file destroy.go to handle the destruction of Terraform components with backend lifecycle considerations.
- Implement methods to manage the destruction of components based on the backend type (kubernetes, s3, azurerm, local).
- Add tests in destroy_test.go to validate the behavior of the new destruction logic, including scenarios for backend migration, error handling, and component-specific destruction.
