+++ cmd/destroy_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"os"
 	"strings"
 	"testing"
 
@@ -20,6 +19,7 @@ import (
 	"github.com/windsorcli/cli/pkg/runtime"
 	"github.com/windsorcli/cli/pkg/runtime/config"
 	"github.com/windsorcli/cli/pkg/runtime/shell"
+	"github.com/windsorcli/cli/pkg/runtime/tools"
 )
 
 // =============================================================================
@@ -32,6 +32,7 @@ type DestroyMocks struct {
 	BlueprintHandler  *blueprint.MockBlueprintHandler
 	TerraformStack    *terraforminfra.MockStack
 	KubernetesManager *kubernetes.MockKubernetesManager
+	ToolsManager      *tools.MockToolsManager
 	Runtime           *runtime.Runtime
 	TmpDir            string
 }
@@ -114,6 +115,7 @@ func setupDestroyTest(t *testing.T, opts ...*SetupOptions) *DestroyMocks {
 		BlueprintHandler:  mockBlueprintHandler,
 		TerraformStack:    mockTerraformStack,
 		KubernetesManager: mockKubernetesManager,
+		ToolsManager:      baseMocks.ToolsManager,
 		Runtime:           rt,
 		TmpDir:            tmpDir,
 	}
@@ -565,408 +567,6 @@ func TestDestroyTerraformCmd(t *testing.T) {
 		}
 	})
 
-	t.Run("DestroysNonBackendThenMigratesAndDestroysBackend", func(t *testing.T) {
-		// Given a blueprint that declares both a "backend" and a non-backend
-		// terraform component, and the configured backend is non-local (s3), the
-		// new symmetric flow must: destroy non-backend components first against the
-		// live remote backend (excludeIDs="backend"), then pin backend.type=local,
-		// migrate just the backend component's state, destroy it, and restore the
-		// configured backend on defer. The old "migrate everything to local first"
-		// flow is gone — the rest of the bucket's state stays remote until the
-		// non-backend layers are torn down.
-		mocks := setupDestroyTest(t)
-		bpWithBackend := &blueprintv1alpha1.Blueprint{
-			Metadata: blueprintv1alpha1.Metadata{Name: "test"},
-			TerraformComponents: []blueprintv1alpha1.TerraformComponent{
-				{Path: "backend"},
-				{Path: "cluster"},
-			},
-		}
-		mocks.BlueprintHandler.GenerateFunc = func() *blueprintv1alpha1.Blueprint { return bpWithBackend }
-
-		mockCH := mocks.ConfigHandler.(*config.MockConfigHandler)
-		mockCH.GetStringFunc = func(key string, defaultValue ...string) string {
-			if key == "terraform.backend.type" {
-				return "s3"
-			}
-			if len(defaultValue) > 0 {
-				return defaultValue[0]
-			}
-			return ""
-		}
-
-		var ops []string
-		mockCH.SetFunc = func(key string, value any) error {
-			if key == "terraform.backend.type" {
-				ops = append(ops, fmt.Sprintf("set:%v", value))
-			}
-			return nil
-		}
-		mocks.TerraformStack.DestroyAllFunc = func(bp *blueprintv1alpha1.Blueprint, excludeIDs ...string) ([]string, error) {
-			ops = append(ops, fmt.Sprintf("destroyAll:exclude=%v", excludeIDs))
-			return nil, nil
-		}
-		mocks.TerraformStack.MigrateComponentStateFunc = func(bp *blueprintv1alpha1.Blueprint, componentID string) error {
-			ops = append(ops, fmt.Sprintf("migrate:%s", componentID))
-			return nil
-		}
-		mocks.TerraformStack.DestroyFunc = func(bp *blueprintv1alpha1.Blueprint, componentID string) (bool, error) {
-			ops = append(ops, fmt.Sprintf("destroy:%s", componentID))
-			return false, nil
-		}
-
-		proj := newDestroyProject(mocks)
-		cmd := createTestDestroyTerraformCmd()
-		ctx := context.WithValue(context.Background(), projectOverridesKey, proj)
-		cmd.SetArgs([]string{"--confirm=test-context"})
-		cmd.SetContext(ctx)
-		if err := cmd.Execute(); err != nil {
-			t.Fatalf("Expected no error, got %v", err)
-		}
-
-		expected := []string{
-			"destroyAll:exclude=[backend]",
-			"set:local",
-			"migrate:backend",
-			"destroy:backend",
-			"set:s3",
-		}
-		if len(ops) != len(expected) {
-			t.Fatalf("Expected %d ops %v, got %d %v", len(expected), expected, len(ops), ops)
-		}
-		for i, want := range expected {
-			if ops[i] != want {
-				t.Errorf("op %d: got %q, want %q (full: %v)", i, ops[i], want, ops)
-			}
-		}
-	})
-
-	t.Run("BackendMigrationFailureAbortsBackendDestroyAndRestoresBackend", func(t *testing.T) {
-		// Given the backend component's state migration fails (e.g. remote backend
-		// unreachable), the backend's destroy must not run — destroy against a
-		// half-migrated state would corrupt the bucket teardown. Non-backend
-		// components have already been successfully destroyed at this point, so
-		// the surfaced error names the backend migration failure specifically; the
-		// configured backend is restored via defer for subsequent operations.
-		mocks := setupDestroyTest(t)
-		bpWithBackend := &blueprintv1alpha1.Blueprint{
-			Metadata: blueprintv1alpha1.Metadata{Name: "test"},
-			TerraformComponents: []blueprintv1alpha1.TerraformComponent{
-				{Path: "backend"},
-				{Path: "cluster"},
-			},
-		}
-		mocks.BlueprintHandler.GenerateFunc = func() *blueprintv1alpha1.Blueprint { return bpWithBackend }
-
-		mockCH := mocks.ConfigHandler.(*config.MockConfigHandler)
-		mockCH.GetStringFunc = func(key string, defaultValue ...string) string {
-			if key == "terraform.backend.type" {
-				return "s3"
-			}
-			if len(defaultValue) > 0 {
-				return defaultValue[0]
-			}
-			return ""
-		}
-
-		var ops []string
-		mockCH.SetFunc = func(key string, value any) error {
-			if key == "terraform.backend.type" {
-				ops = append(ops, fmt.Sprintf("set:%v", value))
-			}
-			return nil
-		}
-		mocks.TerraformStack.DestroyAllFunc = func(bp *blueprintv1alpha1.Blueprint, excludeIDs ...string) ([]string, error) {
-			ops = append(ops, "destroyAll")
-			return nil, nil
-		}
-		mocks.TerraformStack.MigrateComponentStateFunc = func(bp *blueprintv1alpha1.Blueprint, componentID string) error {
-			ops = append(ops, "migrate-fail")
-			return fmt.Errorf("remote backend unreachable")
-		}
-		destroyCalled := false
-		mocks.TerraformStack.DestroyFunc = func(bp *blueprintv1alpha1.Blueprint, componentID string) (bool, error) {
-			destroyCalled = true
-			return false, nil
-		}
-
-		proj := newDestroyProject(mocks)
-		cmd := createTestDestroyTerraformCmd()
-		ctx := context.WithValue(context.Background(), projectOverridesKey, proj)
-		cmd.SetArgs([]string{"--confirm=test-context"})
-		cmd.SetContext(ctx)
-		err := cmd.Execute()
-
-		if err == nil {
-			t.Fatal("Expected migration error to surface, got nil")
-		}
-		if !strings.Contains(err.Error(), "remote backend unreachable") {
-			t.Errorf("Expected underlying migration cause in surfaced message, got %v", err)
-		}
-		if destroyCalled {
-			t.Error("Backend Destroy must not run after MigrateComponentState fails")
-		}
-		expected := []string{"destroyAll", "set:local", "migrate-fail", "set:s3"}
-		if len(ops) != len(expected) {
-			t.Fatalf("Expected %d ops %v, got %d %v", len(expected), expected, len(ops), ops)
-		}
-		for i, want := range expected {
-			if ops[i] != want {
-				t.Errorf("op %d: got %q, want %q (full: %v)", i, ops[i], want, ops)
-			}
-		}
-	})
-
-	t.Run("BackendRestoreFailureEmitsStderrWarning", func(t *testing.T) {
-		// Given the deferred restore (ch.Set with the original backend value) fails
-		// after a successful backend destroy, the error must surface on stderr so
-		// the operator notices that subsequent commands in the same process will
-		// see backend.type stuck on "local". Destroy itself has already succeeded,
-		// so the command exits zero — but silent restore failure would be a
-		// debugging black hole.
-		mocks := setupDestroyTest(t)
-		bpWithBackend := &blueprintv1alpha1.Blueprint{
-			Metadata: blueprintv1alpha1.Metadata{Name: "test"},
-			TerraformComponents: []blueprintv1alpha1.TerraformComponent{
-				{Path: "backend"},
-				{Path: "cluster"},
-			},
-		}
-		mocks.BlueprintHandler.GenerateFunc = func() *blueprintv1alpha1.Blueprint { return bpWithBackend }
-		mockCH := mocks.ConfigHandler.(*config.MockConfigHandler)
-		mockCH.GetStringFunc = func(key string, defaultValue ...string) string {
-			if key == "terraform.backend.type" {
-				return "s3"
-			}
-			if len(defaultValue) > 0 {
-				return defaultValue[0]
-			}
-			return ""
-		}
-		mockCH.SetFunc = func(key string, value any) error {
-			if key == "terraform.backend.type" && value == "s3" {
-				return fmt.Errorf("mock restore failure")
-			}
-			return nil
-		}
-
-		r, w, pipeErr := os.Pipe()
-		if pipeErr != nil {
-			t.Fatalf("Pipe failed: %v", pipeErr)
-		}
-		origStderr := os.Stderr
-		os.Stderr = w
-		defer func() { os.Stderr = origStderr }()
-
-		proj := newDestroyProject(mocks)
-		cmd := createTestDestroyTerraformCmd()
-		ctx := context.WithValue(context.Background(), projectOverridesKey, proj)
-		cmd.SetArgs([]string{"--confirm=test-context"})
-		cmd.SetContext(ctx)
-		execErr := cmd.Execute()
-
-		w.Close()
-		stderrBytes, _ := io.ReadAll(r)
-		stderrOutput := string(stderrBytes)
-
-		if execErr != nil {
-			t.Fatalf("Expected destroy to succeed despite restore failure, got %v", execErr)
-		}
-		if !strings.Contains(stderrOutput, "failed to restore terraform.backend.type") {
-			t.Errorf("Expected stderr warning about restore failure, got: %q", stderrOutput)
-		}
-		if !strings.Contains(stderrOutput, "mock restore failure") {
-			t.Errorf("Expected stderr warning to include underlying cause, got: %q", stderrOutput)
-		}
-	})
-
-	t.Run("SkipsBackendDanceWhenNoBackendComponent", func(t *testing.T) {
-		// Given a blueprint with no backend component, the destroy flow collapses
-		// to a single DestroyAllTerraform pass with no migration dance. This is the
-		// path for blueprints that reference an out-of-band remote backend.
-		mocks := setupDestroyTest(t)
-		mockCH := mocks.ConfigHandler.(*config.MockConfigHandler)
-		mockCH.GetStringFunc = func(key string, defaultValue ...string) string {
-			if key == "terraform.backend.type" {
-				return "s3"
-			}
-			if len(defaultValue) > 0 {
-				return defaultValue[0]
-			}
-			return ""
-		}
-
-		var ops []string
-		mockCH.SetFunc = func(key string, value any) error {
-			if key == "terraform.backend.type" {
-				ops = append(ops, fmt.Sprintf("set:%v", value))
-			}
-			return nil
-		}
-		var seenExclude []string
-		mocks.TerraformStack.DestroyAllFunc = func(bp *blueprintv1alpha1.Blueprint, excludeIDs ...string) ([]string, error) {
-			seenExclude = excludeIDs
-			ops = append(ops, "destroyAll")
-			return nil, nil
-		}
-		migrateCalled := false
-		mocks.TerraformStack.MigrateComponentStateFunc = func(bp *blueprintv1alpha1.Blueprint, componentID string) error {
-			migrateCalled = true
-			return nil
-		}
-
-		proj := newDestroyProject(mocks)
-		cmd := createTestDestroyTerraformCmd()
-		ctx := context.WithValue(context.Background(), projectOverridesKey, proj)
-		cmd.SetArgs([]string{"--confirm=test-context"})
-		cmd.SetContext(ctx)
-		if err := cmd.Execute(); err != nil {
-			t.Fatalf("Expected no error, got %v", err)
-		}
-
-		if migrateCalled {
-			t.Error("Expected MigrateComponentState NOT to be called when no backend component is declared")
-		}
-		if len(seenExclude) != 0 {
-			t.Errorf("Expected no excludes when no backend component, got %v", seenExclude)
-		}
-		if len(ops) != 1 || ops[0] != "destroyAll" {
-			t.Errorf("Expected single destroyAll op, got %v", ops)
-		}
-	})
-
-	t.Run("KubernetesBackendRunsFullCycleDestroy", func(t *testing.T) {
-		// Given a kubernetes-configured backend, the destroy flow takes the full-
-		// cycle path (mirror image of bootstrap's full-cycle): pin backend.type to
-		// local, MigrateState pulls every component's state from the cluster's
-		// Secrets to local files, DestroyAll tears everything down in reverse
-		// against local state, restore on defer. The per-component dance with
-		// excludeIDs must NOT fire — kubernetes can't peel the backend off because
-		// the cluster IS the backend, and once the cluster is going away every
-		// component's state has to be local already.
-		mocks := setupDestroyTest(t)
-		mockCH := mocks.ConfigHandler.(*config.MockConfigHandler)
-		mockCH.GetStringFunc = func(key string, defaultValue ...string) string {
-			if key == "terraform.backend.type" {
-				return "kubernetes"
-			}
-			if len(defaultValue) > 0 {
-				return defaultValue[0]
-			}
-			return ""
-		}
-
-		var ops []string
-		mockCH.SetFunc = func(key string, value any) error {
-			if key == "terraform.backend.type" {
-				ops = append(ops, fmt.Sprintf("set:%v", value))
-			}
-			return nil
-		}
-		mocks.TerraformStack.MigrateStateFunc = func(bp *blueprintv1alpha1.Blueprint) ([]string, error) {
-			ops = append(ops, "migrate-all")
-			return nil, nil
-		}
-		var seenExclude []string
-		mocks.TerraformStack.DestroyAllFunc = func(bp *blueprintv1alpha1.Blueprint, excludeIDs ...string) ([]string, error) {
-			seenExclude = excludeIDs
-			ops = append(ops, "destroyAll")
-			return nil, nil
-		}
-		migrateComponentCalled := false
-		mocks.TerraformStack.MigrateComponentStateFunc = func(bp *blueprintv1alpha1.Blueprint, componentID string) error {
-			migrateComponentCalled = true
-			return nil
-		}
-
-		proj := newDestroyProject(mocks)
-		cmd := createTestDestroyTerraformCmd()
-		ctx := context.WithValue(context.Background(), projectOverridesKey, proj)
-		cmd.SetArgs([]string{"--confirm=test-context"})
-		cmd.SetContext(ctx)
-		if err := cmd.Execute(); err != nil {
-			t.Fatalf("Expected no error, got %v", err)
-		}
-
-		expected := []string{"set:local", "migrate-all", "destroyAll", "set:kubernetes"}
-		if len(ops) != len(expected) {
-			t.Fatalf("Expected %d ops %v, got %d %v", len(expected), expected, len(ops), ops)
-		}
-		for i, want := range expected {
-			if ops[i] != want {
-				t.Errorf("op %d: got %q, want %q (full: %v)", i, ops[i], want, ops)
-			}
-		}
-		if len(seenExclude) != 0 {
-			t.Errorf("Expected no excludeIDs on the kubernetes path, got %v", seenExclude)
-		}
-		if migrateComponentCalled {
-			t.Error("Expected per-component MigrateComponentState NOT to be called on the kubernetes path")
-		}
-	})
-
-	t.Run("KubernetesBackendMigrationFailureAbortsDestroy", func(t *testing.T) {
-		// Given the kubernetes full-cycle destroy's pre-destroy state migration
-		// fails (e.g. the cluster is unreachable or the kubernetes provider
-		// rejects auth), DestroyAll must not run — destroying against an
-		// inconsistent local state would partially tear down resources whose state
-		// terraform doesn't track. The configured backend must still be restored
-		// via defer for any subsequent operations in the same process.
-		mocks := setupDestroyTest(t)
-		mockCH := mocks.ConfigHandler.(*config.MockConfigHandler)
-		mockCH.GetStringFunc = func(key string, defaultValue ...string) string {
-			if key == "terraform.backend.type" {
-				return "kubernetes"
-			}
-			if len(defaultValue) > 0 {
-				return defaultValue[0]
-			}
-			return ""
-		}
-
-		var ops []string
-		mockCH.SetFunc = func(key string, value any) error {
-			if key == "terraform.backend.type" {
-				ops = append(ops, fmt.Sprintf("set:%v", value))
-			}
-			return nil
-		}
-		mocks.TerraformStack.MigrateStateFunc = func(bp *blueprintv1alpha1.Blueprint) ([]string, error) {
-			ops = append(ops, "migrate-fail")
-			return nil, fmt.Errorf("cluster unreachable")
-		}
-		destroyCalled := false
-		mocks.TerraformStack.DestroyAllFunc = func(bp *blueprintv1alpha1.Blueprint, excludeIDs ...string) ([]string, error) {
-			destroyCalled = true
-			return nil, nil
-		}
-
-		proj := newDestroyProject(mocks)
-		cmd := createTestDestroyTerraformCmd()
-		ctx := context.WithValue(context.Background(), projectOverridesKey, proj)
-		cmd.SetArgs([]string{"--confirm=test-context"})
-		cmd.SetContext(ctx)
-		err := cmd.Execute()
-
-		if err == nil {
-			t.Fatal("Expected migration error to surface, got nil")
-		}
-		if !strings.Contains(err.Error(), "cluster unreachable") {
-			t.Errorf("Expected underlying migration cause in surfaced message, got %v", err)
-		}
-		if destroyCalled {
-			t.Error("DestroyAll must not run after MigrateState fails on the kubernetes path")
-		}
-		expected := []string{"set:local", "migrate-fail", "set:kubernetes"}
-		if len(ops) != len(expected) {
-			t.Fatalf("Expected %d ops %v, got %d %v", len(expected), expected, len(ops), ops)
-		}
-		for i, want := range expected {
-			if ops[i] != want {
-				t.Errorf("op %d: got %q, want %q (full: %v)", i, ops[i], want, ops)
-			}
-		}
-	})
 
 	t.Run("CheckAuthFailureBlocksDestroyAllTerraformBeforeStateMigration", func(t *testing.T) {
 		// Given expired credentials, destroy terraform (no arg) must fail at preflight rather
